+++ src/main/java/fll/web/Welcome.java
@@ -9,8 +9,10 @@ package fll.web;
 import java.util.Formatter;
 import java.util.List;
 
+import fll.UserImages;
 import jakarta.servlet.ServletContext;
 import jakarta.servlet.jsp.JspWriter;
+import jakarta.servlet.jsp.PageContext;
 
 /**
  * Helpers for welcome.jsp.
@@ -18,6 +20,26 @@ import jakarta.servlet.jsp.JspWriter;
 public final class Welcome {
 
   private Welcome() {
+  }
+
+  /**
+   * Name of file in {@link UserImages#getImagesPath()} with the logo for the
+   * partner.
+   */
+  public static final String PARTNER_LOGO_FILENAME = "partner_logo.jpg";
+
+  /**
+   * Name of file in {@link UserImages#getImagesPath()} with the logo for FLL.
+   */
+  public static final String FLL_LOGO_FILENAME = "fll_logo.jpg";
+
+  /**
+   * @param page used to set variables for the page
+   */
+  public static void populateContext(final PageContext page) {
+    page.setAttribute("partner_logo",
+                      String.format("%s/%s", UserImages.getImagesPath().getFileName(), PARTNER_LOGO_FILENAME));
+    page.setAttribute("fll_logo", String.format("%s/%s", UserImages.getImagesPath().getFileName(), FLL_LOGO_FILENAME));
   }
 
   // private static final int MAX_NUM_LOGOS_PER_COLUMN = 6;