+++ Changes.md
@@ -1,3 +1,6 @@
+* #1120 - Add support for changing the FLL logo on the subjective rubrics
+* #1119 - Add support for changing the FLL logo on the welcome page
+* #1121 - Add support for changing the partner logo on the welcome page
 * #1128 - Refresh the performance score team selection page when scores are entered
 * #1135 - Allow editing of the award group order on the awards script editor page
 * #1126 - Reload scoreboard if a score is edited to be lower than it was previously